Move namespaces to avoid clash with class name
